refactor(modbus): extract device config value-types into value_types.rs
Stage 2 of #89. Moves Scalar/ValueType/AccessCfg/EndianCfg/AlignmentCfg,
parse_bitmask, and the From<*Cfg> codec conversions out of device.rs; re-exported
so config::device::* paths are unchanged. device.rs 883->690.
